requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
bleach==5.0.1
Django==4.1.5
djangorestframework==3.14.0
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.4.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
webencodings==0.5.1
